# docs/swagger-v1.json
{
    "openapi":  "3.0.4",
    "info":  {
                 "title":  "SmartTaskManagement.API",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/auth/register":  {
                                             "post":  {
                                                          "tags":  [
                                                                       "Auth"
                                                                   ],
                                                          "summary":  "Creates a new user account with the default Team Member role.\r\nDoes not return tokens — use M:SmartTaskManagement.API.Controllers.AuthController.Login(SmartTaskManagement.Application.Authentication.Dtos.LoginRequestDto,System.Threading.CancellationToken) afterwards.",
                                                          "description":  "The password must meet the backend complexity rules (minimum 8 characters, upper, lower, digit, special).",
                                                          "requestBody":  {
                                                                              "description":  "Registration input.",
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/RegisterRequestDto"
                                                                                                                                  }
                                                                                                                   },
                                                                                              "text/json":  {
                                                                                                                "schema":  {
                                                                                                                               "$ref":  "#/components/schemas/RegisterRequestDto"
                                                                                                                           }
                                                                                                            },
                                                                                              "application/*+json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/RegisterRequestDto"
                                                                                                                                    }
                                                                                                                     }
                                                                                          }
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "OK"
                                                                                    }
                                                                        }
                                                      }
                                         },
                  "/api/auth/login":  {
                                          "post":  {
                                                       "tags":  [
                                                                    "Auth"
                                                                ],
                                                       "summary":  "Authenticates a user and returns access and refresh tokens.",
                                                       "requestBody":  {
                                                                           "description":  "Login credentials.",
                                                                           "content":  {
                                                                                           "application/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/LoginRequestDto"
                                                                                                                               }
                                                                                                                },
                                                                                           "text/json":  {
                                                                                                             "schema":  {
                                                                                                                            "$ref":  "#/components/schemas/LoginRequestDto"
                                                                                                                        }
                                                                                                         },
                                                                                           "application/*+json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/LoginRequestDto"
                                                                                                                                 }
                                                                                                                  }
                                                                                       }
                                                                       },
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "OK"
                                                                                 }
                                                                     }
                                                   }
                                      },
                  "/api/auth/refresh":  {
                                            "post":  {
                                                         "tags":  [
                                                                      "Auth"
                                                                  ],
                                                         "summary":  "Rotates an existing refresh token pair and returns a new access token\r\nand refresh token. The presented refresh token is revoked.",
                                                         "requestBody":  {
                                                                             "description":  "Refresh token input.",
                                                                             "content":  {
                                                                                             "application/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/RefreshRequestDto"
                                                                                                                                 }
                                                                                                                  },
                                                                                             "text/json":  {
                                                                                                               "schema":  {
                                                                                                                              "$ref":  "#/components/schemas/RefreshRequestDto"
                                                                                                                          }
                                                                                                           },
                                                                                             "application/*+json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/RefreshRequestDto"
                                                                                                                                   }
                                                                                                                    }
                                                                                         }
                                                                         },
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK"
                                                                                   }
                                                                       }
                                                     }
                                        },
                  "/api/auth/logout":  {
                                           "post":  {
                                                        "tags":  [
                                                                     "Auth"
                                                                 ],
                                                        "summary":  "Revokes the supplied refresh token so it can no longer be exchanged.",
                                                        "requestBody":  {
                                                                            "description":  "Logout input containing the refresh token.",
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/LogoutRequestDto"
                                                                                                                                }
                                                                                                                 },
                                                                                            "text/json":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/LogoutRequestDto"
                                                                                                                         }
                                                                                                          },
                                                                                            "application/*+json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/LogoutRequestDto"
                                                                                                                                  }
                                                                                                                   }
                                                                                        }
                                                                        },
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "OK"
                                                                                  }
                                                                      }
                                                    }
                                       },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "LoginRequestDto":  {
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "email":  {
                                                                                                "type":  "string",
                                                                                                "nullable":  true
                                                                                            },
                                                                                  "password":  {
                                                                                                   "type":  "string",
                                                                                                   "nullable":  true
                                                                                               }
                                                                              },
                                                               "additionalProperties":  false
                                                           },
                                       "LogoutRequestDto":  {
                                                                "type":  "object",
                                                                "properties":  {
                                                                                   "refreshToken":  {
                                                                                                        "type":  "string",
                                                                                                        "nullable":  true
                                                                                                    }
                                                                               },
                                                                "additionalProperties":  false
                                                            },
...
                                       "RefreshRequestDto":  {
                                                                 "type":  "object",
                                                                 "properties":  {
                                                                                    "refreshToken":  {
                                                                                                         "type":  "string",
                                                                                                         "nullable":  true
                                                                                                     }
                                                                                },
                                                                 "additionalProperties":  false
                                                             },
                                       "RegisterRequestDto":  {
                                                                  "type":  "object",
                                                                  "properties":  {
                                                                                     "email":  {
                                                                                                   "type":  "string",
                                                                                                   "nullable":  true
                                                                                               },
                                                                                     "password":  {
                                                                                                      "type":  "string",
                                                                                                      "nullable":  true
                                                                                                  },
                                                                                     "fullName":  {
                                                                                                      "type":  "string",
                                                                                                      "nullable":  true
                                                                                                  }
                                                                                 },
                                                                  "additionalProperties":  false
                                                              },
...
                                   },
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "http",
                                                              "description":  "Enter the JWT access token (without the \u0027Bearer \u0027 prefix).",
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT"
                                                          }
                                           }
                   },
    "security":  [
                     {
                         "Bearer":  [

                                    ]
                     }
                 ],
...
}